Refactor ClassInfoReference
