+++ src/main/java/com/btk5h/skriptmirror/util/ClassInfoReference.java
@@ -48,52 +48,4 @@ public class ClassInfoReference {
     return plural;
   }
 
-  public static Expression<ClassInfoReference> getFromClassInfoExpression(Expression<ClassInfo<?>> expression) {
-    ClassInfoReference parsedReference = SkriptUtil.getClassInfoReference(expression);
-    return new SimpleExpression<ClassInfoReference>() {
-
-      @Override
-      public boolean init(Expression<?>[] expressions, int matchedPattern, Kleenean isDelayed, SkriptParser.ParseResult parseResult) {
-        return expression.init(expressions, matchedPattern, isDelayed, parseResult);
-      }
-
-      @Override
-      protected ClassInfoReference[] get(Event event) {
-        if (parsedReference != null) {
-          return new ClassInfoReference[] { parsedReference };
-        } else if (isSingle()) {
-          ClassInfo<?> classInfo = expression.getSingle(event);
-          if (classInfo == null) {
-            return new ClassInfoReference[0];
-          }
-          return new ClassInfoReference[] { new ClassInfoReference(classInfo) };
-        } else {
-          return expression.stream(event)
-              .filter(Objects::nonNull)
-              .map(ClassInfoReference::new)
-              .toArray(ClassInfoReference[]::new);
-        }
-      }
-
-      public Class<? extends ClassInfoReference> getReturnType() {
-        return ClassInfoReference.class;
-      }
-
-      @Override
-      public boolean isSingle() {
-        return expression.isSingle();
-      }
-
-
-      @Override
-      public String toString(@Nullable Event event, boolean debug) {
-        if (debug) {
-          return expression.toString(event, true) + " (wrapped by ClassInfoReference)";
-        }
-        return expression.toString(event, false);
-      }
-
-    };
-  }
-
 }
